Move run directory scripts and input files to GEOS-Chem repository
In this version of GEOS-Chem (12.7.1) run directory files are still stored
in the Unit Tester, but in 13.0.0 all run directory files will be moved to
this location. The CH4 run directory files have been moved to this repository
in anticipation of merging the CH4 analytical inversion updates with 13.0.0,
to keep the CH4 run directory files in sync with the CH4 code, and to
simplify the analytical inversion setup script (setup_ch4_inversion.sh).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>