Remove GEOS-Chem as a Git submodule
The GEOS-Chem source code is no longer a Git submodule within this
repository. This was done to remove confusion but also to prepare for the
update to GEOS-Chem 13.0.0 which utilizes a wrapper and its own Git
submodules. Instead, we now instruct users to manually clone GEOS-Chem
into this directory (and checkout the CH4_Analytical_Inversion branch)
or to point to their own local clone or fork of the GEOS-Chem source code.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>