Remove fields from config.yml that aren't used or should not be changed
The settings for HalfPolar, Levs, and Buffer should not be changed. These
have been removed from config.yml and the appropriate values are now hard-
coded in setup_imi.sh when replacing text in input.geos. The NestedGrid
logical is kept in config.yml for eventually expanding support of global
grids. Also, REGION has been renamed to NestedRegion and comments have
been updated to make it more clear that this field specifies which pre-cropped
met fields should be used.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>