Bugfix/preview estimate (#89)
As mentioned in #87 the IMI Preview does not take into account the area of the inversion domain in its cost estimation scheme. It currently assumes the inversion domain scales according to the number of state vector elements, which neglected changes in the buffer elements. This PR scales according to the total area of the inversion domain as retrieved by the state vector. Fixed in #89.