Feature/updated-clustering approach (#102)
Under the context of this branch, updated the clustering algorithm to address issues with non-contiguous state vector elements.

New layered-kmeans algorithm:
This new algorithm maintains the same spirit of Hannah's original version, but lends itself better to automation and minimizes the issues with contiguity. In essence, it successively layers different k-means generated state vector labels starting from high resolution and progressively reducing the resolution of the next layer, assigning the highest sensitivity labels from each layer to the final state vector.

In more detail, the algorithm basics:

Initialize a state vector with all 0 labels in the region of interest. 0 represents a pixel that is yet to be assigned.

Generate list of cluster pairs based on sensitivities of pixels and avg dofs per element. This attempts to get a reasonable guess at the optimal cluster pairings, but it does not take into account spatial proximity, relying on the subsequent steps to ensure contiguity between clusters.

Use the highest resolution aggregation level in the list of cluster pairs to determine the number of clusters to use for kmeans

Use k-means to cluster all 0 labeled pixels using 3 features for labeling designation (lat, lon, sensitivity) and the number of clusters specified in step 3. This results in a single layer with approximately evenly clustered elements.

Using the state vector elements from step 4, calculate the average sensitivity per grid cell for each element.

Assign labels to our actual statevector. We only assign the x highest sensitivity clusters based on the cluster pair selected in step 3.

Repeat from step 3 using the next highest resolution aggregation level and only using unassigned grid cells (0 values)

Other Notes on this new algorithm:
By default the lowest aggregation level is a 4x5 degree element and the algorithm "saves" the requisite number of elements needed to accomplish this for the final layer to be applied

It is still possible to get some non-contiguous elements, but only for the lower-resolution elements. In practice, I have not seen this happen in my testing.
It is more computationally expensive than the original simplified version because of the (potentially many) iterations of k-means. It takes ~15 minutes on my local computer to cluster CONUS. To address this I have also included an option to use mini-batch-kmeans, which is very similar to kmeans, but much faster for large datasets with a price of slightly less accuracy. This reduces the time to ~3 minutes for generating the CONUS state vector.
Estimated sensitivities are generated in the same manner as is used in the IMI preview, which takes into account both prior emissions and observation density.

---------

Co-authored-by: nicholasbalasus <nicholasbalasus@g.harvard.edu>