v2023-06-BCs, drop TROPOMI albedo filters, and Rocky 8 on Cannon (#130)
Allows the IMI to run on Cannon after the upgrade from CentOS 7 to Rocky 8 (used Bob's new environment files from Update Harvard Cannon environment files and Conda files for RockyLinux #114 but left our imi_env.yml as is).

Removes the albedo filters for TROPOMI data.

Updates the boundary condition generation scripts so that if you generate boundary conditions up until day x, TROPOMI and GC data is used until day x+15 to make our temporal smoothing valid (and to make the BCs more reproducible). I generate out to 20 days past just to be careful (and because we will probably be doing this on a monthly basis).