Feature/dynamic kf clustering (#160)
This implements a dynamic Kalman Filter clustering option into the IMI, which allows users to update a clustered state vector at each iteration. This allows statevector clusters to respond to changes in information content "dynamically" as the number of observations and distribution of emissions change over each iteration.

For calculating the estimated sensitivities in DynamicKFClustering mode I have updated the code to use 1) observations for the current iteration instead of the entire inversion period and 2) the nudge (prior) emissions. This allows the clustering algorithm to take into account updated emissions, while still maintaining some presence of the original prior emissions.

Also includes some small bugfixes to:

- the run_notebooks function previously cd'ed into the incorrect directory when in KalmanMode
- updated i to period_i in kalman.sh and other scripts where it is inherited