Add option for OH optimization to IMI
Additional options to optimize for OH have been added to config.yml. This
includes the switch for turning this option on (OptimizeOH), PriorErrorOH,
and PerturbValueOH. When enabling this option, an additional run will be
added to the jacobian runs. Code has also been added to the inversion
scripts for including OH optimization.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>