Remove land cover threshold as default in make_state_vector_file.py
A threshold of land > 1% and ice < 10% was added when determining land
cover for generation of the the state vector file. This was added for the
global inversions, but we have disabled that as the default option here.
If users want to impose a threshold, they can do so on their own or we
could add this as a future option.

Specifying the threshold and/or code setting state vector to nan or 0 resulted
in netCDF file that could not be properly read by HEMCO, so those updates
have also been reverted.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>