Condense environment file activations (#194)
This is related to running the IMI at Harvard on Cannon. For @laestrada and I, the combination of first sourcing the GEOS-Chem environment file and then activating the conda environment (an alias for micromamba in our case) results in GEOS-Chem not being able to run (./gcclassic: error while loading shared libraries: libnetcdf.so.18: cannot open shared object file: No such file or directory).

This order was introduced in 7c57d91. This PR is a fix that makes sure the Conda environment and GEOS-Chem environment file are only activated a single time each (now in run_imi.sh) and than the order is Conda then GEOS-Chem environment file.