bugfix for defining geoschem grid based on grid cell edges (not centers) (#190)
When defining the domain in geoschem_config.yml, we use the statevector max and min lat/lons based on the grid cell centers. However, geoschem defines the domain based on the grid cell edges. This requires a slight adjustment to the calculation of the geoschem domain: adding/subtracting half the spacing between grid cells, so that the geoschem outputted SpeciesConc domain matches the statevector.

Note that this adjustment only needs to take place in the geoschem_config.yml file during setup of the template run directory.