send an error from run_inversion.sh batch job (#196)
Previously if the run_inversion.sh script failed with an error, the inversion workflow would continue despite the error because run_inversion.sh is run via an sbatch job.

To fix this, on any non-zero exit code this script writes out a hidden error file that the inversion workflow reads and errors out on. This is how we handle other failed sbatch jobs eg. jacobian runs in the IMI.