Add fixes to get model output from correct jacobian run directory
With the capability to now combine jacobian runs into fewer runs the inversion
scripts needed to be updated to look for the model output for each state
vector element in the correct location. This involves passing the number of
runs defined in the config file to the inversion scripts so that they can
determine how many state vector elements are in each jacobian run.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>