fix bug in specification of geoschem domain for some custom domains (#207)
This fixes a bug with the specification of the longitude bounds for the geoschem domain. According to the geoschem [documentation](https://geos-chem.readthedocs.io/en/latest/gcclassic-user-guide/geoschem-config.html#cmdoption-arg-latitude), the northernmost/southernmost grid boxes are half the size of others for both merra2 and geosfp. This only affects latitudes. Previously we were accounting for this for both lat and lon. This removes the correction for longitude.