Feature/point_sources: new point source capabilities and clustering updates (#219)
This pull request includes 2 enhancements:

Enabling use of point source information from IMEO and CarbonMapper in the state vector definition. This also updates the way that SRON plumes are retrieved. Most of the infrastructure to do this is in src/inversion_scripts/plumes.py
Add capability to cluster by country in addition to clustering based on estimated DOFS in state vector aggregation. Most of these updates are in src/components/statevector_component/aggregation.py

* Add CarbonMapper and IMEO as plume data sources. Optionally force native res state vector at plumes from these.

* SRON capability, read locally archived file until scraping fix available. Add plume count and emission rate filter options for forcing native resolution.


---------

Co-authored-by: laestrada <laestrada00@gmail.com>
Co-authored-by: Lucas A Estrada <63303345+laestrada@users.noreply.github.com>