Bugfix to check if retrieval time rounds to outside of the current period (#236)
This addresses the bug in #211 in which retrieval times can be rounded to a time outside of the period of interest due to rounding to the nearest hour. In the TROPOMI operator, an observation within the GC start and end dates can be rounded to be equal to the end date 00 hour, which is problematic because GC does not output for this hour. To avoid this problem, those dates are floor rounded.