Bugfix/allow zero buffer elems (#224)
Previously the IMI required use of buffer elements in the state vector when making a rectilinear state vector. This update allows specification of 0 buffer elements by setting the buffer region == -9999. This PR also enforces a minimum "buffer zone" of 3 grid cells on each side is added to the state vector.