Bugfix/prior emissions: switch to daily precompiled emissions (#239)
This addresses #237.
When a non-kalman-filter inversion runs more than one month, HEMCO emissions files are generated for each month. In places like the preview and visualization notebook, only the first month of emissions are used. So if you ran a 1 year inversion, the emissions reported are just for January.

This update computes the daily emissions and updates them daily during the jacobian and posterior simulations.

I also add logic to compute the time averaged emissions for the inversion period for python scripts.

---------

Co-authored-by: nicholasbalasus <nicholasbalasus@g.harvard.edu>