Remove "preview_run" and instead save IMI preview output in "preview" directory
The preview simulation is no longer run so we do not need an actual run
directory (preview_run). Instead, we now use the HEMCO standalone simulation
in prior_run. This is now cleaned up in the code. Output from the IMI preview
(plots, text) will now simply be saved in a directory called "preview."

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>