Now apply hemispheric OH perturbation using a netCDF mask file defining hemispheres
Applying masks to scale factors does not work in HEMCO as advertised. Instead,
we utilize the same method used to perturb state vector elements to perturb
hemispheric OH. That is, we use a mask file defining hemispheric regions and
apply the perturbations via a txt file matching to the correct index in that
mask file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>