Bugfix/new-env (#264)
With the new conda environment, there are some new warnings in the IMI test case. One is for duplicate dimensions (nvar) which is fixed here.

The other is for NaNs being passed to the uint8 casting, which is also fixed here. This is with respect to the surface_classification variable in TROPOMI files. There are no NaNs in the blended files because QA flag filtering has already occurred.

The posterior was also failing with Infinity in DO_CLOUD_CONVECTION! which was due to NaNs in the scale factors. These NaNs are set to 1.