Rename prior component and run directory to reflect HEMCO standalone simulation and not true prior run
There was some confusion about the use of prior_run and prior.sh which was
used for the HEMCO standalone simulation used to generate total prior emissions
output for use in the preview, jacobian simulations, etc. The prior run in
the inversion actually refers to jacobian run 0000 (i.e. the model run with
zero perturbations applied). To avoid confusion, we now rename the prior_run
directory to hemco_prior_emis. The prior component is now referred to as the
hemco_proior_emis component.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>