Add temporary fix for path to HEMCO prior emissions output in HEMCO_Config.rc
The path to the total prior emissions is hardcoded in the default HEMCO
configuration file obtained from GEOS-Chem to use ../../prior_run. This
directory was renamed in the previous commit. For now, we can use sed to
fix the directory name in HEMCO_Config.rc in the template run directory.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>