Add fixes to *invert.py scripts to properly determine number of OH elements
In lognormal_inverty.py it was automatically assumed that the number of OH
elements is 1 when setting OptimizeOH: true. This is only the case for regional
simulations. For global simulations, OH_element_num should be 2.

In invert.py, a fix was needed to select the right element in K when using
OH optimization in regional vs global simulations. In addition, the logicals
for BC and OH optimization have been renamed in this script for consistency
with lognormal_invert.py.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>