Add option for validating IMI prior and posterior simulations against NOAA ObsPack data (#329)
* Add option for validating prior and posterior simulation against NOAA ObsPack

A new option has been added to the configuration file `DoObsPackValidation`
which will turn on the ObsPack diagnostic option in GEOS-Chem for the
prior (jacobian run 0000) and posterior simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>

* Add sample notebook for comparing GEOS-Chem against NOAA Obspack data

The script NOAA_ObsPack_MBL_compare.ipnyb compares NOAA ObsPack marine
boundary later sites to GEOS-Chem ObsPack diagnostic output. This script
was provided by Megan He (Harvard).

Some further thought needs to go into including ObsPack vs GEOS-Chem
plots by default in the standard visualization notebook (when the
DoObsPackValidation option is set to true in the configuration file). Not
all IMI simulations/domains will necessarily include ObsPack output even when
the option is enabled. For example, this is the case with the out-of-the-box
Permian example.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>

* Add ObsPack option to documentation

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>

---------

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>