Update sample configuration file for 12km runs and remove custom lines used for 12km testing
In several places, the ConfigPath was changed for testing 12km runs. This has
been restored to look in the InversionPath. The code to copy the config file
to InversionPath has also been restored.

In statevector.sh, the LandCoverSuffix for 12km runs has been updated to
remove a user-specific path and point to the IMERGE landcover file now
available in the HEMCO/CH4/v2025-03 data path.

Finally, the config.harvard-cannon.12km.yml file has been updated for
consistency with the latest IMI development branch.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>