Update configuration files to use blended TROPOMI observations and boundary conditions (#354)
The blended TROPOMI+GOSAT product described in Balasus et al. (2023) is now
recommended as the default product for use in the IMI. This option was originally
introduced in IMI 2.0 as documented in Estrada et al. (2025). The default configuration
file (config.yml) and envs/Harvard-Cannon configuration files have been updated to
reflect this recommendation. The configuration files have also been updated
to use the blended boundary condition files for consistency with the observational
dataset used.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>