Fix date string in geoschem_config.yml for 12km runs
The default date string in template_run/geoschem_config.yml for CH4 runs
was 20190101. This is not the case for 12km runs though because fields
at that resolution are not available until mid-2021. When users set up
a 12km GEOS-Chem run directory, the date in geoschem_config.yml is set to
20230101 by default.

The code in template.sh has been updated to look for the correct date string
based on grid resolution and replace it with the start and end dates as
defined in the IMI config file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>