Update HEMCO prior emission run to save out one diagnostic file instead of daily files
The HEMCO prior run is needed to obtain average emissions over the inversion
period. Daily output it not needed but was specified in the HEMCO config file.
The `DiagnFreq` option has been changed from Daily back to End so that
only one HEMCO diagnostic file is saved for the prior emission run.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>