Bug fix: Remove time dimension from state vector on IMI preview's estimate_averaging_kernel
To avoid a "tuple index out of range" error when creating the buffer_mask
and referencing the dimensions of state_vector_labels, the time dimension
should first be removed. The time dimension was added to the state vector
netCDF file for compatibility with GCHP (and MAPL).

Also update make_state_vector_file.py to use the correct long_name. Otherwise
it is saved out as 'StateVector:long_name = "Fraction of land in grid box"'.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>