Change timecycle flags for restart files to allow for restart files missing other carbon species
Now that we are using the carbon simulation, GEOS-Chem and HEMCO expect to find
all species (chemical and advected) included in that simulation in the
restart file. The default restart files (taken from TROPOMI-corrected boundary
condition files) only contain CH4 and not the remaining carbon species (e.g. CO2,
CO) so HEMCO currently crashes saying it cannot find those fields in the
restart file. To avoid this, we change the cycle flag in HEMCO_Config.rc from
EFYO to CYS to tell HEMCO to skip a species if it's not found and set it to
default background concentrations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>